requirements.txt
Flask~=3.1.1
Flask-SQLAlchemy~=3.1.1
Flask-Migrate~=4.0.5
pytest~=8.3.1
pytest-cov~=4.0.0
pytest-mock~=3.12.0
pytest-asyncio~=0.23.0
# Bootstrap is included via static files or CDN
# Microsoft Graph API
msgraph-sdk~=1.28.0
# Xero API (official SDK)
xero-python~=8.1.0
# OpenAI API
openai~=1.28.0
# Google Maps Directions API
googlemaps~=4.10.0
# For encrypted fields in SQLAlchemy models
sqlalchemy-utils~=0.41.2
requests-oauthlib~=2.0.0
python-dotenv~=1.1.0
opentelemetry-api~=1.33.0
opentelemetry-sdk~=1.33.0
opentelemetry-instrumentation-flask~=0.54b0
opentelemetry-exporter-otlp~=1.33.0
Flask-Login~=0.6.3
msal~=1.28.0
Flask-APScheduler~=1.13.1
pytz~=2024.1
apprise~=1.7.5
typer[all]~=0.16.0  # For CLI support
rich~=14.0.0
nest_asyncio~=1.6.0